backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-generativeai==0.8.5
firebase-admin==6.2.0
elevenlabs==0.2.27
python-dotenv==1.0.0